Removed preprocess option #1510
